djelastic/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
Django==4.2
django-elasticsearch-dsl==8.0
django-redis==5.4.0
djangorestframework==3.14.0
elastic-transport==8.12.0
gunicorn==21.2.0
kombu==5.3.5
packaging==24.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
